mon_travail/security_ai_app/requirements1.txt
streamlit
numpy
torch
torchvision
ultralytics
tensorflow-cpu
opencv-python-headless
Pillow